requirements.txt
absl-py==1.0.0
apache-beam==2.38.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
astunparse==1.6.3
attrs==21.4.0
avro-python3==1.10.2
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.0
cachetools==5.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cloudpickle==2.0.0
colorama==0.4.4
commonmark==0.9.1
contextlib2==21.6.0
crcmod==1.7
cycler==0.11.0
Cython==0.29.28
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.1.1
dm-tree==0.1.7
docopt==0.6.2
docutils==0.18.1
entrypoints==0.4
executing==0.8.3
fastavro==1.4.11
fastjsonschema==2.15.3
flatbuffers==2.0
fonttools==4.33.3
gast==0.5.3
gin-config==0.5.0
google-api-core==2.7.3
google-api-python-client==2.47.0
google-auth==2.6.6
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.56.1
grpcio==1.46.1
h5py==3.6.0
hdfs==2.7.0
httplib2==0.19.1
idna==3.3
importlib-metadata==4.11.3
ipykernel==6.13.0
ipython==8.3.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jsonschema==4.5.1
jupyter==1.0.0
jupyter-client==7.3.1
jupyter-console==6.4.3
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
kaggle==1.5.12
keras==2.8.0
Keras-Preprocessing==1.1.2
keyring==23.5.0
kiwisolver==1.4.2
libclang==14.0.1
lvis==0.5.3
lxml==4.8.0
Markdown==3.3.7
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mistune==0.8.4
nbclient==0.6.3
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
notebook==6.4.11
numpy==1.22.3
oauth2client==4.1.3
oauthlib==3.2.0
object-detection==0.1
opencv-python==4.5.5.64
opt-einsum==3.3.0
orjson==3.6.8
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.1.0
pkginfo==1.8.2
portalocker==2.4.0
prometheus-client==0.14.1
promise==2.3
prompt-toolkit==3.0.29
proto-plus==1.20.3
protobuf==3.20.1
psutil==5.9.0
pure-eval==0.2.2
py-cpuinfo==8.0.0
pyarrow==6.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.4
pycparser==2.21
pydot==1.4.2
Pygments==2.12.0
pymongo==3.12.3
pyparsing==2.4.7
pyrsistent==0.18.1
python-dateutil==2.8.2
python-slugify==6.1.2
pytz==2022.1
pywin32==304
pywin32-ctypes==0.2.0
pywinpty==2.0.5
PyYAML==5.4.1
pyzmq==22.3.0
qtconsole==5.3.0
QtPy==2.1.0
readme-renderer==35.0
regex==2022.4.24
requests==2.27.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rfc3986==2.0.0
rich==12.4.1
rsa==4.8
sacrebleu==2.0.0
scikit-learn==1.1.0
scipy==1.8.0
Send2Trash==1.8.0
sentencepiece==0.1.96
seqeval==1.2.2
six==1.16.0
soupsieve==2.3.2.post1
stack-data==0.2.0
tabulate==0.8.9
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-addons==0.16.1
tensorflow-datasets==4.5.2
tensorflow-hub==0.12.0
tensorflow-io==0.25.0
tensorflow-io-gcs-filesystem==0.25.0
tensorflow-metadata==1.8.0
tensorflow-model-optimization==0.7.2
tensorflow-text==2.8.2
termcolor==1.1.0
terminado==0.13.3
text-unidecode==1.3
tf-estimator-nightly==2.8.0.dev2021122109
tf-models-official==2.8.0
tf-slim==1.1.0
threadpoolctl==3.1.0
tinycss2==1.1.1
tornado==6.1
tqdm==4.64.0
traitlets==5.2.0
twine==4.0.0
typeguard==2.13.3
typing_extensions==4.2.0
uritemplate==4.1.1
urllib3==1.26.9
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.1.2
widgetsnbextension==3.6.0
wrapt==1.14.1
zipp==3.8.0
